Split tweetQuote into tweetQuoteText and tweetQuoteImage
