+++ functions/tweetQuoteText.yml
@@ -1,11 +1,11 @@
 
-tweetQuote:
-  handler: src/tweetQuote.handler
-  description: Tweet a random quote.
+tweetQuoteText:
+  handler: src/tweetQuoteText.handler
+  description: Tweet quote as text.
   environment:
     TWITTER_CONSUMER_KEY: ${ssm:nietzsche-twitter-consumer-key}
     TWITTER_CONSUMER_SECRET: ${ssm:nietzsche-twitter-consumer-secret}
     TWITTER_ACCESS_TOKEN: ${ssm:nietzsche-twitter-access-token}
     TWITTER_ACCESS_TOKEN_SECRET: ${ssm:nietzsche-twitter-access-token-secret}
-  memorySize: 512
-  timeout: 300
+  memorySize: 128
+  timeout: 20
